requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
pydantic>=1.9.0
python-multipart>=0.0.5
aiofiles>=0.8.0
numpy
scikit-learn
requests>=2.31.0
beautifulsoup4>=4.12.2
openai
matplotlib
prometheus_client
# Text extraction dependencies
python-docx>=0.8.11
PyPDF2>=3.0.0
# Testing dependencies
pytest>=7.3.1
pytest-asyncio>=0.21.0
